+++ NEWS.md
@@ -12,6 +12,10 @@
 
 -   During the process of translating messages, many messages were improved and made more consistent.
 
+-   New Function `ResidLm()` that returns the residuals of a linear fit.
+
+-   New function `Detrend()` that, you guessed it, returns a (linearly) detrended version of the input vector.
+
 ## Bugfixes
 
 -   `seasonally()` result will be on the 15th of the centre month of each season instead of on the 1st. This makes the date more representative of the time span and also solves a bug in which dates on the 31st would return `NA`. This is a **potentially breaking change**.